[nfc] Resolve incompatible type comparison warnings
The recent changes from PR #723 cause a handful of comparisons in the
code to trigger warnings. None of them are actual problems, and this
patch disambiguates the comparisons to remove the warning.

+++ lib/Targets/AIETargetSimulationFiles.cpp
@@ -380,7 +380,8 @@ mlir::LogicalResult AIETranslateGraphXPE(mlir::ModuleOp module,
       int row = tileOp.rowIndex();
 
       // NOTE: row == 0 assumes shim always row 0
-      if (row == 0 || row > targetOp.getTargetModel().getNumMemTileRows())
+      if (row == 0 ||
+          row > static_cast<int>(targetOp.getTargetModel().getNumMemTileRows()))
         continue; // Skip regular tiles (handled above)
 
       output << "      <MEM name=\"MEM(" << std::to_string(col) << ", "
